Move some code into testing-specific module
