+++ messenger/shared/qgraph_compiler.py
@@ -138,90 +138,3 @@ class EdgeReference():
             return self._filters
         else:
             return ''
-
-
-def cypher_query_fragment_match(qgraph, max_connectivity=-1):
-    """Generate a Cypher query fragment to match the nodes and edges that correspond to a question.
-
-    This is used internally for cypher_query_answer_map and cypher_query_knowledge_graph
-
-    Returns the query fragment as a string.
-    """
-    nodes, edges = qgraph['nodes'], qgraph['edges']
-
-    # generate internal node and edge variable names
-    node_references = {n['id']: NodeReference(n) for n in nodes}
-    edge_references = [EdgeReference(e) for e in edges]
-
-    match_strings = []
-
-    # match orphaned nodes
-    def flatten(l):
-        return [e for sl in l for e in sl]
-    all_nodes = {n['id'] for n in nodes}
-    all_referenced_nodes = set(flatten([[e['source_id'], e['target_id']] for e in edges]))
-    orphaned_nodes = all_nodes - all_referenced_nodes
-    for n in orphaned_nodes:
-        match_strings.append(f"MATCH ({node_references[n]})")
-        match_strings[-1] += node_references[n].extras
-        if node_references[n].filters:
-            match_strings.append("WHERE " + node_references[n].filters)
-
-    # match edges
-    for e, eref in zip(edges, edge_references):
-        source_node = node_references[e['source_id']]
-        target_node = node_references[e['target_id']]
-        match_strings.append(f"MATCH ({source_node}){eref}({target_node})")
-        match_strings[-1] += source_node.extras + target_node.extras
-        filters = [f'({c})' for c in [source_node.filters, target_node.filters, eref.filters] if c]
-        if max_connectivity > -1:
-            filters.append(f"(size( ({target_node})-[]-() ) < {max_connectivity})")
-        if filters:
-            match_strings.append("WHERE " + " AND ".join(filters))
-
-    match_string = ' '.join(match_strings)
-
-    return match_string
-
-
-def cypher_query_answer_map(qgraph, **kwargs):
-    """Generate a Cypher query to extract the answer maps for a question.
-
-    Returns the query as a string.
-    """
-    clauses = []
-
-    match_string = cypher_query_fragment_match(qgraph, max_connectivity=kwargs.pop('max_connectivity', -1))
-    if match_string:
-        clauses.append(match_string)
-
-    nodes, edges = qgraph['nodes'], qgraph['edges']
-
-    # generate internal node and edge variable names
-    node_names = [f"{n['id']}" for n in nodes]
-    edge_names = [f"{e['id']}" for e in edges]
-
-    # deal with sets
-    node_id_accessor = [f"collect(DISTINCT {n['id']}.id) AS {n['id']}" if 'set' in n and n['set'] else f"[{n['id']}.id] AS {n['id']}" for n in nodes]
-    if kwargs.get('relationship_id', 'property') == 'internal':
-        edge_id_accessor = [f"collect(DISTINCT toString(id({e['id']}))) AS {e['id']}" for e in edges]
-    else:
-        edge_id_accessor = [f"collect(DISTINCT {e['id']}.id) AS {e['id']}" for e in edges]
-    if node_id_accessor or edge_id_accessor:
-        with_string = f"WITH {', '.join(node_id_accessor+edge_id_accessor)}"
-        clauses.append(with_string)
-
-    # add bound fields and return map
-    node_dicts = [f"[ni IN {n} | {{qg_id:'{n}', kg_id:ni}}]" for n in node_names]
-    edge_dicts = [f"[ei IN {e} | {{qg_id:'{e}', kg_id:ei}}]" for e in edge_names]
-    answer_return_string = f"RETURN {' + '.join(node_dicts) or '[]'} AS nodes, {' + '.join(edge_dicts) or '[]'} AS edges"
-    clauses.append(answer_return_string)
-
-    # return answer maps matching query
-    query_string = ' '.join(clauses)
-    if 'skip' in kwargs:
-        query_string += f' SKIP {kwargs["skip"]}'
-    if 'limit' in kwargs:
-        query_string += f' LIMIT {kwargs["limit"]}'
-
-    return query_string
