+++ tests/test_novelty.py
@@ -7,7 +7,7 @@ from reasoner_pydantic import Request
 
 from messenger.modules.weight_novelty import query as novelty
 from messenger.modules.weight_correctness import query as correctness
-from fixtures import to_weight
+from .fixtures import to_weight
 
 
 def test_weight(to_weight):
