+++ tests/test_score.py
@@ -3,7 +3,7 @@
 # ^^^ this stuff happens because of the incredible way we do pytest fixtures
 import json
 from messenger.modules.score import query as score
-from fixtures import weighted2
+from .fixtures import weighted2
 
 
 def test_score(weighted2):
